backend/requirements.txt
Flask
Flask-JWT-Extended
psycopg2-binary
bcrypt
python-dotenv